Item Tiers 1-B: reshape item instances to the 15-byte modifier POD; Compatibility 0.8.0 (#40)
item_instance_data attribute portion is now the locked spec 12 POD
(custom_made, tier, modifiers[4]{type,value,value2}, enchant_bonus);
positional AttributePrefixType/SecondaryEffectType enums deleted with
ItemAttributes.h. All roll/apply/price/enchant/display code speaks
unified modifier IDs; the one legacy->unified translation table lives in
ModifierIds.h and translates at the three boundaries (crafted-recipe
nibbles, GM tester bitmask, legacy-keyed attribute-type config rows).
CItem accessors are legacy views over modifiers[0]/[1]; wire/DB keep
legacy field names carrying unified IDs until 1-C/1-E. The [16] lookup
arrays die (unified [256] interim until the 1-D catalog); nibble-era
value-15 roll clamp deleted (drift #7). Compatibility 0.7.0 -> 0.8.0 in
version.cfg - the single bump of the effort; mixed pairs rejected
pre-auth. Windows All green x2, zero compiler warnings; /simplify pass
applied (GM dialog derives multipliers through the translation table).

+++ Sources/Server/AccountSqliteStore.cpp
@@ -1744,9 +1744,9 @@ bool SaveCharacterSnapshot(sqlite3* db, const CClient* client)
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_list[i]->m_instance.special_effect_value3) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_list[i]->m_instance.cur_durability) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_list[i]->is_custom_made()) == SQLITE_OK);
-        ok &= (sqlite3_bind_int(stmt, col++, static_cast<int>(client->m_item_list[i]->get_prefix_type())) == SQLITE_OK);
+        ok &= (sqlite3_bind_int(stmt, col++, client->m_item_list[i]->get_prefix_type()) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_list[i]->get_prefix_value()) == SQLITE_OK);
-        ok &= (sqlite3_bind_int(stmt, col++, static_cast<int>(client->m_item_list[i]->get_secondary_type())) == SQLITE_OK);
+        ok &= (sqlite3_bind_int(stmt, col++, client->m_item_list[i]->get_secondary_type()) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_list[i]->get_secondary_value()) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_list[i]->get_enchant_bonus()) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_pos_list[i].x) == SQLITE_OK);
@@ -1807,9 +1807,9 @@ bool SaveCharacterSnapshot(sqlite3* db, const CClient* client)
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_in_bank_list[i]->m_instance.special_effect_value3) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_in_bank_list[i]->m_instance.cur_durability) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_in_bank_list[i]->is_custom_made()) == SQLITE_OK);
-        ok &= (sqlite3_bind_int(stmt, col++, static_cast<int>(client->m_item_in_bank_list[i]->get_prefix_type())) == SQLITE_OK);
+        ok &= (sqlite3_bind_int(stmt, col++, client->m_item_in_bank_list[i]->get_prefix_type()) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_in_bank_list[i]->get_prefix_value()) == SQLITE_OK);
-        ok &= (sqlite3_bind_int(stmt, col++, static_cast<int>(client->m_item_in_bank_list[i]->get_secondary_type())) == SQLITE_OK);
+        ok &= (sqlite3_bind_int(stmt, col++, client->m_item_in_bank_list[i]->get_secondary_type()) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_in_bank_list[i]->get_secondary_value()) == SQLITE_OK);
         ok &= (sqlite3_bind_int(stmt, col++, client->m_item_in_bank_list[i]->get_enchant_bonus()) == SQLITE_OK);
